Mock and functional-scoped test grouping (fixtures)
