+++ tests/cli/functional/conftest.py
@@ -14,20 +14,40 @@ GNU Affero General Public License for more details.
 You should have received a copy of the GNU Affero General Public License
 along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
 """
-
+import os
 import pytest
+from eth_account import Account
 
 from nucypher.blockchain.economics import EconomicsFactory
 from nucypher.blockchain.eth.agents import ContractAgency
-from nucypher.blockchain.eth.interfaces import BlockchainInterface
-from tests.mock.agents import MockWorkLockAgent, FAKE_RECEIPT, MockContractAgency
+from nucypher.blockchain.eth.interfaces import BlockchainInterface, BlockchainInterfaceFactory
+from nucypher.blockchain.eth.registry import InMemoryContractRegistry
+from nucypher.config.characters import UrsulaConfiguration
+from nucypher.utilities.sandbox.blockchain import MockBlockchain
+from tests.cli.functional.test_ursula_local_keystore_cli_functionality import NUMBER_OF_MOCK_ACCOUNTS, \
+    KEYFILE_NAME_TEMPLATE
+from tests.fixtures import _make_testerchain
+from tests.mock.agents import FAKE_RECEIPT, MockContractAgency
 
+
+@pytest.fixture(scope='module', autouse=True)
+def mock_testerchain() -> MockBlockchain:
+    BlockchainInterfaceFactory._interfaces = dict()
+    testerchain = _make_testerchain(mock_backend=True)
+    yield testerchain
+    
 
 @pytest.fixture(scope='module', autouse=True)
 def mock_interface(module_mocker):
     mock_transaction_sender = module_mocker.patch.object(BlockchainInterface, 'sign_and_broadcast_transaction')
     mock_transaction_sender.return_value = FAKE_RECEIPT
     return mock_transaction_sender
+
+
+@pytest.fixture(scope='session')
+def test_registry():
+    registry = InMemoryContractRegistry()
+    return registry
 
 
 @pytest.fixture(scope='module', autouse=True)
@@ -50,8 +70,29 @@ def mock_contract_agency(module_mocker, token_economics):
     ContractAgency.get_agent_by_contract_name = get_agent_by_name
 
 
-@pytest.fixture(scope='function')
-def mock_worklock_agent(mock_testerchain, token_economics):
-    mock_agent = MockWorkLockAgent()
-    yield mock_agent
-    mock_agent.reset()
+@pytest.fixture(scope='module')
+def mock_accounts():
+    accounts = dict()
+    for i in range(NUMBER_OF_MOCK_ACCOUNTS):
+        account = Account.create()
+        filename = KEYFILE_NAME_TEMPLATE.format(month=i+1, address=account.address)
+        accounts[filename] = account
+    return accounts
+
+
+@pytest.fixture(scope='module')
+def worker_account(mock_accounts, mock_testerchain):
+    account = list(mock_accounts.values())[0]
+    return account
+
+
+@pytest.fixture(scope='module')
+def worker_address(worker_account):
+    address = worker_account.address
+    return address
+
+
+@pytest.fixture(scope='module')
+def custom_config_filepath(custom_filepath):
+    filepath = os.path.join(custom_filepath, UrsulaConfiguration.generate_filename())
+    return filepath
